+++ src/Game/Text/HelpText.de.sjson.lua
@@ -7,7 +7,17 @@ local order = {
 local newData = {
   {
     Id = "ModsNikkelMCrossroadSingingSirensMusicPlayer_Song_Scylla1a_Lyrics",
+    DisplayName = "Lied der Sirenen (Humming)",
+    Description = "Ein Lied von einem Musik-Trio aus dem Okeanos, das ein großes, begeistertes Publikum anlocken soll."
+  },
+  {
+    Id = "ModsNikkelMCrossroadSingingSirensMusicPlayer_Song_Scylla1a_Lyrics_Actual",
     DisplayName = "Lied der Sirenen (Lyrics)",
+    Description = "Ein Lied von einem Musik-Trio aus dem Okeanos, das ein großes, begeistertes Publikum anlocken soll."
+  },
+  {
+    Id = "ModsNikkelMCrossroadSingingSirensMusicPlayer_Song_Scylla1a_Lyrics_Hummed",
+    DisplayName = "Lied der Sirenen (Hummed Melody)",
     Description = "Ein Lied von einem Musik-Trio aus dem Okeanos, das ein großes, begeistertes Publikum anlocken soll."
   },
   {
@@ -17,7 +27,17 @@ local newData = {
   },
   {
     Id = "ModsNikkelMCrossroadSingingSirensMusicPlayer_Song_Scylla2a_Lyrics",
+    DisplayName = "Lied der Tiefe (Humming)",
+    Description = "Ein Lied zu Ehren der auf See verschollenen Matrosen von einem Musik-Trio aus dem Okeanos."
+  },
+  {
+    Id = "ModsNikkelMCrossroadSingingSirensMusicPlayer_Song_Scylla2a_Lyrics_Actual",
     DisplayName = "Lied der Tiefe (Lyrics)",
+    Description = "Ein Lied zu Ehren der auf See verschollenen Matrosen von einem Musik-Trio aus dem Okeanos."
+  },
+  {
+    Id = "ModsNikkelMCrossroadSingingSirensMusicPlayer_Song_Scylla2a_Lyrics_Hummed",
+    DisplayName = "Lied der Tiefe (Hummed Melody)",
     Description = "Ein Lied zu Ehren der auf See verschollenen Matrosen von einem Musik-Trio aus dem Okeanos."
   },
   {
@@ -27,7 +47,17 @@ local newData = {
   },
   {
     Id = "ModsNikkelMCrossroadSingingSirensMusicPlayer_Song_Scylla3a_Lyrics",
+    DisplayName = "Lied der Inseln (Humming)",
+    Description = "Ein Lied über die Seefahrt von einem Musik-Trio aus dem Okeanos."
+  },
+  {
+    Id = "ModsNikkelMCrossroadSingingSirensMusicPlayer_Song_Scylla3a_Lyrics_Actual",
     DisplayName = "Lied der Inseln (Lyrics)",
+    Description = "Ein Lied über die Seefahrt von einem Musik-Trio aus dem Okeanos."
+  },
+  {
+    Id = "ModsNikkelMCrossroadSingingSirensMusicPlayer_Song_Scylla3a_Lyrics_Hummed",
+    DisplayName = "Lied der Inseln (Hummed Melody)",
     Description = "Ein Lied über die Seefahrt von einem Musik-Trio aus dem Okeanos."
   },
   {
@@ -37,6 +67,11 @@ local newData = {
   },
   {
     Id = "ModsNikkelMCrossroadSingingSirensMusicPlayer_Song_Scylla4a_Lyrics",
+    DisplayName = "Lied der Perlen (Humming)",
+    Description = "Ein Lied über die Schätze des Meeres von einem Musik-Trio aus dem Okeanos."
+  },
+  {
+    Id = "ModsNikkelMCrossroadSingingSirensMusicPlayer_Song_Scylla4a_Lyrics_Actual",
     DisplayName = "Lied der Perlen (Lyrics)",
     Description = "Ein Lied über die Schätze des Meeres von einem Musik-Trio aus dem Okeanos."
   },