+++ src/Game/Text/HelpText.el.sjson.lua
@@ -7,7 +7,19 @@ local order = {
 local newData = {
   {
     Id = "ModsNikkelMCrossroadSingingSirensMusicPlayer_Song_Scylla1a_Lyrics",
+    DisplayName = "Τραγούδι των Σειρήνων (Humming)",
+    Description =
+    "Μουσική που αποσκοπεί να δελεάσει και να μαγέψει ένα μεγάλο κοινό, με εκτέλεση από ένα μουσικό τρίο που κατοικοεδρεύει στον Ωκεανό."
+  },
+  {
+    Id = "ModsNikkelMCrossroadSingingSirensMusicPlayer_Song_Scylla1a_Lyrics_Actual",
     DisplayName = "Τραγούδι των Σειρήνων (Lyrics)",
+    Description =
+    "Μουσική που αποσκοπεί να δελεάσει και να μαγέψει ένα μεγάλο κοινό, με εκτέλεση από ένα μουσικό τρίο που κατοικοεδρεύει στον Ωκεανό."
+  },
+  {
+    Id = "ModsNikkelMCrossroadSingingSirensMusicPlayer_Song_Scylla1a_Lyrics_Hummed",
+    DisplayName = "Τραγούδι των Σειρήνων (Hummed Melody)",
     Description =
     "Μουσική που αποσκοπεί να δελεάσει και να μαγέψει ένα μεγάλο κοινό, με εκτέλεση από ένα μουσικό τρίο που κατοικοεδρεύει στον Ωκεανό."
   },
@@ -19,7 +31,19 @@ local newData = {
   },
   {
     Id = "ModsNikkelMCrossroadSingingSirensMusicPlayer_Song_Scylla2a_Lyrics",
+    DisplayName = "Τραγούδι του Βυθού (Humming)",
+    Description =
+    "Μουσική προς τιμήν των ναυτικών που χάθηκαν για πάντα στη θάλασσα, με εκτέλεση από ένα μουσικό τρίο που κατοικοεδρεύει στον Ωκεανό."
+  },
+  {
+    Id = "ModsNikkelMCrossroadSingingSirensMusicPlayer_Song_Scylla2a_Lyrics_Actual",
     DisplayName = "Τραγούδι του Βυθού (Lyrics)",
+    Description =
+    "Μουσική προς τιμήν των ναυτικών που χάθηκαν για πάντα στη θάλασσα, με εκτέλεση από ένα μουσικό τρίο που κατοικοεδρεύει στον Ωκεανό."
+  },
+  {
+    Id = "ModsNikkelMCrossroadSingingSirensMusicPlayer_Song_Scylla2a_Lyrics_Hummed",
+    DisplayName = "Τραγούδι του Βυθού (Hummed Melody)",
     Description =
     "Μουσική προς τιμήν των ναυτικών που χάθηκαν για πάντα στη θάλασσα, με εκτέλεση από ένα μουσικό τρίο που κατοικοεδρεύει στον Ωκεανό."
   },
@@ -31,7 +55,19 @@ local newData = {
   },
   {
     Id = "ModsNikkelMCrossroadSingingSirensMusicPlayer_Song_Scylla3a_Lyrics",
+    DisplayName = "Τραγούδι των Νησιών (Humming)",
+    Description =
+    "Μουσική προς τιμήν της ναυτικής εξερεύνησης, με εκτέλεση από ένα μουσικό τρίο που κατοικοεδρεύει στον Ωκεανό."
+  },
+  {
+    Id = "ModsNikkelMCrossroadSingingSirensMusicPlayer_Song_Scylla3a_Lyrics_Actual",
     DisplayName = "Τραγούδι των Νησιών (Lyrics)",
+    Description =
+    "Μουσική προς τιμήν της ναυτικής εξερεύνησης, με εκτέλεση από ένα μουσικό τρίο που κατοικοεδρεύει στον Ωκεανό."
+  },
+  {
+    Id = "ModsNikkelMCrossroadSingingSirensMusicPlayer_Song_Scylla3a_Lyrics_Hummed",
+    DisplayName = "Τραγούδι των Νησιών (Hummed Melody)",
     Description =
     "Μουσική προς τιμήν της ναυτικής εξερεύνησης, με εκτέλεση από ένα μουσικό τρίο που κατοικοεδρεύει στον Ωκεανό."
   },
@@ -43,6 +79,12 @@ local newData = {
   },
   {
     Id = "ModsNikkelMCrossroadSingingSirensMusicPlayer_Song_Scylla4a_Lyrics",
+    DisplayName = "Τραγούδι των Μαργαριταριών (Humming)",
+    Description =
+    "Μουσική με θεματολογία τους θησαυρούς της θάλασσας, με εκτέλεση από ένα μουσικό τρίο που κατοικοεδρεύει στον Ωκεανό."
+  },
+  {
+    Id = "ModsNikkelMCrossroadSingingSirensMusicPlayer_Song_Scylla4a_Lyrics_Actual",
     DisplayName = "Τραγούδι των Μαργαριταριών (Lyrics)",
     Description =
     "Μουσική με θεματολογία τους θησαυρούς της θάλασσας, με εκτέλεση από ένα μουσικό τρίο που κατοικοεδρεύει στον Ωκεανό."