+++ src/Game/Text/HelpText.tr.sjson.lua
@@ -7,7 +7,19 @@ local order = {
 local newData = {
   {
     Id = "ModsNikkelMCrossroadSingingSirensMusicPlayer_Song_Scylla1a_Lyrics",
+    DisplayName = "Sirenlerin Şarkısı (Humming)",
+    Description =
+    "Okeanos çıkışlı müzik üçlüsünün, geniş ve sadık bir dinleyici kitlesini çekmeyi amaçlayan tema müziği."
+  },
+  {
+    Id = "ModsNikkelMCrossroadSingingSirensMusicPlayer_Song_Scylla1a_Lyrics_Actual",
     DisplayName = "Sirenlerin Şarkısı (Lyrics)",
+    Description =
+    "Okeanos çıkışlı müzik üçlüsünün, geniş ve sadık bir dinleyici kitlesini çekmeyi amaçlayan tema müziği."
+  },
+  {
+    Id = "ModsNikkelMCrossroadSingingSirensMusicPlayer_Song_Scylla1a_Lyrics_Hummed",
+    DisplayName = "Sirenlerin Şarkısı (Hummed Melody)",
     Description =
     "Okeanos çıkışlı müzik üçlüsünün, geniş ve sadık bir dinleyici kitlesini çekmeyi amaçlayan tema müziği."
   },
@@ -18,7 +30,17 @@ local newData = {
   },
   {
     Id = "ModsNikkelMCrossroadSingingSirensMusicPlayer_Song_Scylla2a_Lyrics",
+    DisplayName = "Derinlerin Şarkısı (Humming)",
+    Description = "Okeanos çıkışlı müzik üçlüsünün, denize açılıp kaybolan denizcilere ithaf ettiği tema müziği."
+  },
+  {
+    Id = "ModsNikkelMCrossroadSingingSirensMusicPlayer_Song_Scylla2a_Lyrics_Actual",
     DisplayName = "Derinlerin Şarkısı (Lyrics)",
+    Description = "Okeanos çıkışlı müzik üçlüsünün, denize açılıp kaybolan denizcilere ithaf ettiği tema müziği."
+  },
+  {
+    Id = "ModsNikkelMCrossroadSingingSirensMusicPlayer_Song_Scylla2a_Lyrics_Hummed",
+    DisplayName = "Derinlerin Şarkısı (Hummed Melody)",
     Description = "Okeanos çıkışlı müzik üçlüsünün, denize açılıp kaybolan denizcilere ithaf ettiği tema müziği."
   },
   {
@@ -29,7 +51,17 @@ local newData = {
   },
   {
     Id = "ModsNikkelMCrossroadSingingSirensMusicPlayer_Song_Scylla3a_Lyrics",
+    DisplayName = "Adaların Şarkısı (Humming)",
+    Description = "Okeanos çıkışlı müzik üçlüsünün, deniz keşiflerine ithaf ettiği tema müziği."
+  },
+  {
+    Id = "ModsNikkelMCrossroadSingingSirensMusicPlayer_Song_Scylla3a_Lyrics_Actual",
     DisplayName = "Adaların Şarkısı (Lyrics)",
+    Description = "Okeanos çıkışlı müzik üçlüsünün, deniz keşiflerine ithaf ettiği tema müziği."
+  },
+  {
+    Id = "ModsNikkelMCrossroadSingingSirensMusicPlayer_Song_Scylla3a_Lyrics_Hummed",
+    DisplayName = "Adaların Şarkısı (Hummed Melody)",
     Description = "Okeanos çıkışlı müzik üçlüsünün, deniz keşiflerine ithaf ettiği tema müziği."
   },
   {
@@ -39,6 +71,11 @@ local newData = {
   },
   {
     Id = "ModsNikkelMCrossroadSingingSirensMusicPlayer_Song_Scylla4a_Lyrics",
+    DisplayName = "İncilerin Şarkısı (Humming)",
+    Description = "Okeanos çıkışlı müzik üçlüsünün, denizlerin zenginliklerini anlattığı tema müziği."
+  },
+  {
+    Id = "ModsNikkelMCrossroadSingingSirensMusicPlayer_Song_Scylla4a_Lyrics_Actual",
     DisplayName = "İncilerin Şarkısı (Lyrics)",
     Description = "Okeanos çıkışlı müzik üçlüsünün, denizlerin zenginliklerini anlattığı tema müziği."
   },