+++ src/Scripts/MusicPlayerData.lua
@@ -1,14 +1,33 @@
+-- Section 0 of a biome ambience track is the humming, Section 1 is the singing
+-- The value of a new song in this table is the song after which the new one should be inserted
 mod.ModdedCrossroadSongs = {
 	ModsNikkelMCrossroadSingingSirensMusicPlayer_Song_Scylla1a_Lyrics = "Song_Scylla1a",
+	ModsNikkelMCrossroadSingingSirensMusicPlayer_Song_Scylla1a_Lyrics_Hummed =
+	"ModsNikkelMCrossroadSingingSirensMusicPlayer_Song_Scylla1a_Lyrics",
+	ModsNikkelMCrossroadSingingSirensMusicPlayer_Song_Scylla1a_Lyrics_Actual =
+	"ModsNikkelMCrossroadSingingSirensMusicPlayer_Song_Scylla1a_Lyrics_Hummed",
 	ModsNikkelMCrossroadSingingSirensMusicPlayer_Song_Scylla1b_Lyrics = "Song_Scylla1b",
 	ModsNikkelMCrossroadSingingSirensMusicPlayer_Song_Scylla2a_Lyrics = "Song_Scylla2a",
+	ModsNikkelMCrossroadSingingSirensMusicPlayer_Song_Scylla2a_Lyrics_Hummed =
+	"ModsNikkelMCrossroadSingingSirensMusicPlayer_Song_Scylla2a_Lyrics",
+	ModsNikkelMCrossroadSingingSirensMusicPlayer_Song_Scylla2a_Lyrics_Actual =
+	"ModsNikkelMCrossroadSingingSirensMusicPlayer_Song_Scylla2a_Lyrics_Hummed",
 	ModsNikkelMCrossroadSingingSirensMusicPlayer_Song_Scylla2b_Lyrics = "Song_Scylla2b",
 	ModsNikkelMCrossroadSingingSirensMusicPlayer_Song_Scylla3a_Lyrics = "Song_Scylla3a",
+	ModsNikkelMCrossroadSingingSirensMusicPlayer_Song_Scylla3a_Lyrics_Hummed =
+	"ModsNikkelMCrossroadSingingSirensMusicPlayer_Song_Scylla3a_Lyrics",
+	ModsNikkelMCrossroadSingingSirensMusicPlayer_Song_Scylla3a_Lyrics_Actual =
+	"ModsNikkelMCrossroadSingingSirensMusicPlayer_Song_Scylla3a_Lyrics_Hummed",
 	ModsNikkelMCrossroadSingingSirensMusicPlayer_Song_Scylla3b_Lyrics = "Song_Scylla3b",
 	ModsNikkelMCrossroadSingingSirensMusicPlayer_Song_Scylla4a_Lyrics = "Song_Scylla4a",
+	ModsNikkelMCrossroadSingingSirensMusicPlayer_Song_Scylla4a_Lyrics_Actual =
+	"ModsNikkelMCrossroadSingingSirensMusicPlayer_Song_Scylla4a_Lyrics",
 	ModsNikkelMCrossroadSingingSirensMusicPlayer_Song_Scylla4b_Lyrics = "Song_Scylla4b",
 }
 
+-- The Lyrics versions are actually Humming, this was incorrectly named in the initial version of the mod
+-- Lyrics_Actual is the actual lyrical/singing version
+-- Lyrics_Hummed is the lyrical section, but hummed
 local songWorldUpgradeData = {
 	ModsNikkelMCrossroadSingingSirensMusicPlayer_Song_Scylla1a_Lyrics = {
 		Name = "ModsNikkelMCrossroadSingingSirensMusicPlayer_Song_Scylla1a_Lyrics",
@@ -23,6 +42,38 @@ local songWorldUpgradeData = {
 		},
 		Cost = {
 			CosmeticsPoints = 400,
+			MixerGBoss = 1,
+		},
+	},
+	ModsNikkelMCrossroadSingingSirensMusicPlayer_Song_Scylla1a_Lyrics_Actual = {
+		Name = "ModsNikkelMCrossroadSingingSirensMusicPlayer_Song_Scylla1a_Lyrics_Actual",
+		InheritFrom = { "DefaultSongItem" },
+		TrackName = "/Music/IrisMusicScylla1_MC",
+		ModsNikkelMCrossroadSingingSirens_MusicSection = 1,
+
+		GameStateRequirements = {
+			{
+				PathTrue = { "GameState", "WorldUpgradesAdded", "Song_Scylla1a" },
+			},
+		},
+		Cost = {
+			CosmeticsPoints = 550,
+			MixerGBoss = 1,
+		},
+	},
+	ModsNikkelMCrossroadSingingSirensMusicPlayer_Song_Scylla1a_Lyrics_Hummed = {
+		Name = "ModsNikkelMCrossroadSingingSirensMusicPlayer_Song_Scylla1a_Lyrics_Hummed",
+		InheritFrom = { "DefaultSongItem" },
+		TrackName = "/Music/IrisMusicScylla1_MC",
+		ModsNikkelMCrossroadSingingSirens_MusicSection = 3,
+
+		GameStateRequirements = {
+			{
+				PathTrue = { "GameState", "WorldUpgradesAdded", "ModsNikkelMCrossroadSingingSirensMusicPlayer_Song_Scylla1a_Lyrics_Actual" },
+			},
+		},
+		Cost = {
+			CosmeticsPoints = 550,
 			MixerGBoss = 1,
 		},
 	},
@@ -56,6 +107,38 @@ local songWorldUpgradeData = {
 		},
 		Cost = {
 			CosmeticsPoints = 400,
+			MixerGBoss = 1,
+		},
+	},
+	ModsNikkelMCrossroadSingingSirensMusicPlayer_Song_Scylla2a_Lyrics_Actual = {
+		Name = "ModsNikkelMCrossroadSingingSirensMusicPlayer_Song_Scylla2a_Lyrics_Actual",
+		InheritFrom = { "DefaultSongItem" },
+		TrackName = "/Music/IrisMusicScylla2_MC",
+		ModsNikkelMCrossroadSingingSirens_MusicSection = 1,
+
+		GameStateRequirements = {
+			{
+				PathTrue = { "GameState", "WorldUpgradesAdded", "Song_Scylla2a" },
+			},
+		},
+		Cost = {
+			CosmeticsPoints = 550,
+			MixerGBoss = 1,
+		},
+	},
+	ModsNikkelMCrossroadSingingSirensMusicPlayer_Song_Scylla2a_Lyrics_Hummed = {
+		Name = "ModsNikkelMCrossroadSingingSirensMusicPlayer_Song_Scylla2a_Lyrics_Hummed",
+		InheritFrom = { "DefaultSongItem" },
+		TrackName = "/Music/IrisMusicScylla2_MC",
+		ModsNikkelMCrossroadSingingSirens_MusicSection = 3,
+
+		GameStateRequirements = {
+			{
+				PathTrue = { "GameState", "WorldUpgradesAdded", "ModsNikkelMCrossroadSingingSirensMusicPlayer_Song_Scylla2a_Lyrics_Actual" },
+			},
+		},
+		Cost = {
+			CosmeticsPoints = 550,
 			MixerGBoss = 1,
 		},
 	},
@@ -93,6 +176,38 @@ local songWorldUpgradeData = {
 			MixerGBoss = 1,
 		},
 	},
+	ModsNikkelMCrossroadSingingSirensMusicPlayer_Song_Scylla3a_Lyrics_Actual = {
+		Name = "ModsNikkelMCrossroadSingingSirensMusicPlayer_Song_Scylla3a_Lyrics_Actual",
+		InheritFrom = { "DefaultSongItem" },
+		TrackName = "/Music/IrisMusicScylla3_MC",
+		ModsNikkelMCrossroadSingingSirens_MusicSection = 1,
+
+		GameStateRequirements = {
+			{
+				PathTrue = { "GameState", "WorldUpgradesAdded", "Song_Scylla3a" },
+			},
+		},
+		Cost = {
+			CosmeticsPoints = 750,
+			MixerGBoss = 1,
+		},
+	},
+	ModsNikkelMCrossroadSingingSirensMusicPlayer_Song_Scylla3a_Lyrics_Hummed = {
+		Name = "ModsNikkelMCrossroadSingingSirensMusicPlayer_Song_Scylla3a_Lyrics_Hummed",
+		InheritFrom = { "DefaultSongItem" },
+		TrackName = "/Music/IrisMusicScylla3_MC",
+		ModsNikkelMCrossroadSingingSirens_MusicSection = 3,
+
+		GameStateRequirements = {
+			{
+				PathTrue = { "GameState", "WorldUpgradesAdded", "ModsNikkelMCrossroadSingingSirensMusicPlayer_Song_Scylla3a_Lyrics_Actual" },
+			},
+		},
+		Cost = {
+			CosmeticsPoints = 750,
+			MixerGBoss = 1,
+		},
+	},
 	ModsNikkelMCrossroadSingingSirensMusicPlayer_Song_Scylla3b_Lyrics = {
 		Name = "ModsNikkelMCrossroadSingingSirensMusicPlayer_Song_Scylla3b_Lyrics",
 		InheritFrom = { "DefaultSongItem" },
@@ -127,6 +242,22 @@ local songWorldUpgradeData = {
 			MixerGBoss = 1,
 		},
 	},
+	ModsNikkelMCrossroadSingingSirensMusicPlayer_Song_Scylla4a_Lyrics_Actual = {
+		Name = "ModsNikkelMCrossroadSingingSirensMusicPlayer_Song_Scylla4a_Lyrics_Actual",
+		InheritFrom = { "DefaultSongItem" },
+		TrackName = "/Music/IrisMusicScylla4_MC",
+		ModsNikkelMCrossroadSingingSirens_MusicSection = 1,
+
+		GameStateRequirements = {
+			{
+				PathTrue = { "GameState", "WorldUpgradesAdded", "Song_Scylla4a" },
+			},
+		},
+		Cost = {
+			CosmeticsPoints = 750,
+			MixerGBoss = 1,
+		},
+	},
 	ModsNikkelMCrossroadSingingSirensMusicPlayer_Song_Scylla4b_Lyrics = {
 		Name = "ModsNikkelMCrossroadSingingSirensMusicPlayer_Song_Scylla4b_Lyrics",
 		InheritFrom = { "DefaultSongItem" },
@@ -147,18 +278,29 @@ local songWorldUpgradeData = {
 	},
 }
 
--- Insert the new song right after the original
-for songKey, songValue in pairs(mod.ModdedCrossroadSongs) do
-	local insertIndex = nil
-	for i, existingSong in ipairs(game.ScreenData.MusicPlayer.Songs) do
-		if existingSong == songValue then
-			insertIndex = i
-			break
-		end
-	end
+-- Insert the humming versions after the original, and the lyrical versions after the humming ones
+local songsToInsert = {}
+for songKey, anchorSong in pairs(mod.ModdedCrossroadSongs) do
+	songsToInsert[songKey] = anchorSong
+end
 
-	if insertIndex then
-		table.insert(game.ScreenData.MusicPlayer.Songs, insertIndex + 1, songKey)
+local insertedSong = true
+while insertedSong do
+	insertedSong = false
+	for songKey, anchorSong in pairs(songsToInsert) do
+		local insertIndex = nil
+		for i, existingSong in ipairs(game.ScreenData.MusicPlayer.Songs) do
+			if existingSong == anchorSong then
+				insertIndex = i
+				break
+			end
+		end
+
+		if insertIndex then
+			table.insert(game.ScreenData.MusicPlayer.Songs, insertIndex + 1, songKey)
+			songsToInsert[songKey] = nil
+			insertedSong = true
+		end
 	end
 end